Gateway/requirements.txt
#### `requirements.txt`
```plaintext
fastapi
uvicorn
google-cloud-speech
pydantic
```